fix -Wshorten-64-to-32 warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38197 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/diskimage/fsimage-probe.c
@@ -133,7 +133,7 @@ static int disk_image_check_for_d64(disk_image_t *image)
 
     if (checkimage_errorinfo) {
         fsimage->error_info.map = lib_calloc(1, checkimage_blocks);
-        fsimage->error_info.len = checkimage_blocks;
+        fsimage->error_info.len = (int)checkimage_blocks;
         if (util_fpread(fsimage->fd, fsimage->error_info.map, checkimage_blocks, 256 * checkimage_blocks) < 0) {
             return 0;
         }
